pyproject.toml
[tool.poetry]
name = "PrintingPress"
version = "1.0.0"
description = "A tool to place text and images onto a Pillow image object."
authors = ["hysrx"]

[tool.poetry.dependencies]
python = "^3.6"
Pillow-SIMD = "^7.0.0.post3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
